Add new QVVS Hierarchy Update Modes
This requires a demo and full validation, but the idea is that you can force children to use explicitly set world-space transform properties, either partially or completely. And this will be accounted for directly in the hierarchy update to avoid execution order issues that would otherwise arise.

It's an awesome feature that can only work with QVVS. Lots of gameplay potential with this one.