Start of ITWN group project
This starts the NetForce team's ITWN group project as a dev dungeon.

Along with getting everything set up, this also fixes compound colliders without custom authoring. And it proves the efficacy of Transform Hierarchy Update Modes.

The character controller V1 should probably not be used for any serious project. It is there to serve as a comparison for new and improved techniques.

Next up, some new ergonomic improvement.