Add Rng ShuffleSequence and more
The scene wasn't saved correctly for the last commit, so this one fixes that.

RngSequence.ShuffleSequence() can be used to randomly reorder elements in a NativeArray or INativeList.

I've also started an implementation for DynamicHashMap. Unlike others floating around, this one will properly serialize Entity and BlobAssetReference. That will probably be my muse for the next few commits.