Add GameFlags
As Free Parking is quest-driven with the potential for a quest to be anything, it makes the most sense to implement a global flag system. The idea is that GameFlagAuthoring is converted to a GameFlagHandle during baking. A GameFlagHandle can live anywhere, including inside components and inside blob assets. And a GameFlags collection component on the worldBlackboardEntity provides runtime access for setting, clearing, and querying flags. Flags can be interacted with both in the main world and dev dungeons.

Camera is going to require experimentation with an actual level, and I don't feel comfortable starting it yet. We can just parent it to a player character in the meantime.

Next up, either the interaction scripting runtime, or quests. Or something totally different. Not sure yet.