Add quests and initial interaction script components
Quests are just data definitions. It will be up to the runtime UI systems to interpret them.

Interaction scripting is a little trickier, as there are so many possible things people may want to do, and it can be hard to encapsulate. I'm especially worried about async operations.

I think my motivation is worn out, but if I do get it back, the next step will be the main world pause menu and the various systems that tie into that.