Fix dynamic mesh and blend shape culling and collider size
I knew there was a bug I discovered but never got around to fixing, but I could never remember what that was. Well, I found it again, and it was the loop for blend shapes and dynamic meshes not accounting for the bitmasks. This should probably be backported to 0.8.x at some point.

Also, there's an issue where colliders were using 56 bytes instead of 48 bytes. This was because Triangles were using 36 bytes offset from 16 bytes. I've altered the layout to put the blob asset in the last 8 bytes. All blob collider types are strictly 32 bytes in size (even though some would be fine with 24) offset from 16 bytes. Meanwhile, the primitive collider types are offset from byte 4 so that they have 36 bytes of space to work with.