Fix optimized root bone safety issue and add ParameterClip BlobBuilder extension
Had a little extra time today, so made these two new improvements. There's a new IAspect for getting readonly access to an optimized skeleton's root bone transform, which is typically the root motion delta. This aspect does not conflict with TransformAspect and can be safely used with TransformAspect in an IJobEntity. Mecanim has been updated to use this.

As for the ParameterClip BlobBuilder extension, this allows ParameterClip instances to be baked inside custom blob assets safely.