Convex vs Convex contact manifolds and Kinemation improvements
The last of our 15 non-composite pairs for contact manifolds is written. There's still plenty of cleanup and debugging work left, but this is quite the milestone!

Along with this, there were a couple of improvements to Kinemation. An overallocation issue was fixed with blend shapes. And there's a new method for animation clips to figure out how many loop wraps there were between two unbounded time values.