Fix compound colliders and Calligraphics change filtering
Compound collider inertia tensors are still not correct when extreme scale values in both directions are applied. I'm not sure what the cause of this is. Will need to investigate more. But otherwise things seem to work again.

I did get an issue where an in parameter was being defensively copied causing reading of the wrong blob data. I had to rework the code to do that a little more safely. Very annoying.