Calligraphics Overhaul
Fribur submitted a pull request to rework rich text tag parsing to be closer in functionality and behavior to TextMeshPro. The intent is to port over much of TMP's functionality.

Before all this happens, there were a couple of longstanding issues that needed to get cleaned up. First, obtaining glyphs from the font blob was slow, as it cycled through a bunch of fat objects. Now, there's an accelerated lookup.

Second, support has been added for multiple fonts and materials. While the authoring infrastructure and some of the components were there, the systems and shaders never supported them. Now they do. Although glyph parsing will need to be modified to actually make use of the feature.