LOD Crossfade is working
Finally. The new LOD algorithm is here. Less structural changes, and less position lookups are the big benefits. But also, LOD Crossfade is respected, kinda.

URP 14 has a bug that breaks LOD Crossfade for BRG. I've created a Shader SubGraph in the ShaderLibrary folder of Kinemation which simply needs to be attached indirectly to the vertex output block, and the problem will be fixed for the shader. Hopefully Unity will backport the fix which has been in Unity 6 for a while now.