Optimize DistanceBetweenAll() for box vs triMesh
Two things were slow. First, there was the excessive reallocations of the results buffer. And second, the AABB of the large box was covering the entire mesh, resulting in every triangle being tested with the full GJK algorithm. A new capacity inflation scheme was introduced, and in box vs triangle, an early test of the triangle's AABB in box-space was added, which eliminated most of the GJK evaluations. These changes netted a full 10x performance improvement.