Add GPU-resident text and font tags
Calligraphics got some new features. GPU-resident text trades off culling awareness for persistence. For some text use cases, the glyph data rarely changes and the text is often visible, so keeping it on the GPU is advantageous.

Font switching was pretty trivial to insert. All it took was a new integer index stack in the TextConfiguration and then being able to search for it in the FontMaterialSet.

Lastly, AddComponentsCommandBuffer was added. Being able to force the survival of cleanup components on short-lived entities by putting them on a new entity instead is powerful.