Clean up text configurations
First, I brought in most of Fribur's latest changes. Then I cleaned up the base configuration and made it smaller. ECS chunk memory is too precious. After that, I cleaned up the configuration stack so that it reuses fixed lists in consecutive calls, and also parses rich text a character in advance, which fixes a few bugs.

Next up is to clean up the arithmetic and get this a bit more optimized. And also document the code a little better.